Get rid of the redundant GitHub Actions macOS package builder
This broke at some point and now fails with a syntax error coming out of
setuptools_scm.  This may have to do with setuptools installing a bad version
of that library (which is pulled in via setup_requires).

However, since we have this package building on CircleCI now we don't need to
debug this, we can just drop it.