See if we can fix up CI a little bit
We don't need different versions of nixpkgs anymore.  The flake pins some
version of nixpkgs and that is that.

Also, since it is mostly just deleting stuff, switch to cachix which will do a
better job of caching the Nix pieces (and, otherwise, we no longer have a very
coherent cache key to use).