Simplify the "pkgsForHost" thing
This is really just a simple renaming but the motivation is somewhat
interesting.  We don't *need* a whole "separate" instance of nixpkgs
with a customized "host" system.  That's the *only* instance we need.  Its
build system should just be correct for the system where we're
running (caller's responsibility).  And then use its "host" system
configuration to figure out any cross-compilation necessary.  And that's it.

The code _already_ didn't accept another instance of nixpkgs.  So just name
this one in the conventional way.

Also some more comments.